Refactoring out GlobalContext: import problems resolved and all unit tests passing again
