+++ frontend-react/src/components/Admin/EditReceiverSettings.tsx
@@ -6,7 +6,10 @@ import { RouteComponentProps, useHistory } from "react-router-dom";
 import { ErrorPage } from "../../pages/error/ErrorPage";
 import OrgReceiverSettingsResource from "../../resources/OrgReceiverSettingsResource";
 import { showAlertNotification, showError } from "../AlertNotifications";
-import { getStoredOktaToken, getStoredOrg } from "../GlobalContextProvider";
+import {
+    getStoredOktaToken,
+    getStoredOrg,
+} from "../../contexts/SessionStorageTools";
 import { jsonSortReplacer } from "../../utils/JsonSortReplacer";
 import { CheckFeatureFlag } from "../../pages/misc/FeatureFlags";
