+++ frontend-react/src/resources/AuthResource.ts
@@ -3,7 +3,7 @@ import { Resource } from "@rest-hooks/rest";
 import {
     getStoredOktaToken,
     getStoredOrg,
-} from "../components/GlobalContextProvider";
+} from "../contexts/SessionStorageTools";
 
 export default class AuthResource extends Resource {
     pk(parent?: any, key?: string): string | undefined {
